add 400 error handling + test, also renamed files to organise by table
