+++ app.js
@@ -1,9 +1,12 @@
 const express = require("express");
-const { getCategories } = require("./controller/seed.controller.js");
+const { getCategories } = require("./controllers/categories.controller.js");
 const app = express();
-app.use(express.json());
 
 app.get("/api/categories", getCategories);
+
+app.all("/*", (req, res, next) => {
+  res.status(400).send({ msg: "Bad Request" });
+});
 
 app.use((err, req, res, next) => {
   req.status(500).send({ msg: "Internal Server Error" });
